+++ src/main/java/nextstep/subway/applicaion/PathService.java
@@ -5,7 +5,7 @@ import nextstep.subway.applicaion.dto.StationResponse;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.LineRepository;
 import nextstep.subway.domain.Station;
-import nextstep.subway.utils.PathFinder;
+import nextstep.subway.domain.PathFinder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
